refactor: Refactor kube.rs and move to utils
