+++ rust/stackable-cockpit/src/lib.rs
@@ -2,7 +2,6 @@ pub mod cluster;
 pub mod common;
 pub mod constants;
 pub mod helm;
-pub mod kube;
 pub mod platform;
 pub mod utils;
 pub mod xfer;
